Use CMake policy CMP0135 for CMake >=3.24 to suppress warning
CMake >=3.24 has changed the behaviour of FetchContent_Declare() but the
new behaviour doesn't impact this project. The old behaviour can still
be accessed using the DOWNLOAD_EXTRACT_TIMESTAMP parameter which causes
older CMake versions to fail. So it's easier to simply accept the new
behaviour if it's available to suppress the warning.