Update FetchContent_Populate() for CMake policy CMP0169
CMP0169 has deprecated the use of FetchContent_Populate() with only a
single parameter. Therefore, the URL and URL_HASH parameters are
provided directly to FetchContent_Populate() instead of using
FetchContent_Declare() beforehand.

Note that CMP0169 recommends the use of FetchContent_MakeAvailable()
instead of FetchContent_Populate() but does not prohibit or deprecate
FetchContent_Populate() as a whole. In this particular case the use of
policy CMP0077 to override the MbedTLS build options makes
FetchContent_Populate() preferable.